rename signal_models to model_critters
