+++ .gitignore
@@ -261,7 +261,7 @@ fabric.properties
 # Android studio 3.1+ serialized cache file
 .idea/caches/build_file_checksums.ser
 
-/climatecritters/signal_models/insolation.py
+/climatecritters/model_critters/insolation.py
 /scratchwork/
 
 # # Quarto / quartodoc build outputs
