+++ docs/get-started/concepts.qmd
@@ -123,7 +123,7 @@ Once in Pyleoclim, the full analysis toolkit is available:
 ```python
 import numpy as np
 import climatecritters as cc
-from climatecritters.signal_models import EBM
+from climatecritters.model_critters import EBM
 
 # composable forcing: spin-up then a solar ramp
 forcing = cc.core.Forcing.from_sequence([
