+++ docs/get-started/noise.qmd
@@ -41,7 +41,7 @@ where $\mathrm{d}W$ is a Wiener increment and $g(t, y)$ (the `noise_func` argume
 
 ```python
 import numpy as np
-from climatecritters.signal_models.stommel import Stommel
+from climatecritters.model_critters.stommel import Stommel
 from climatecritters.utils.forcing_utils import create_constant_forcing
 
 model = Stommel(forcing=create_constant_forcing(0.0))
